Управление, смена информации и аватарки сообщества / Поиск по друзьям, группам, глобал

+++ frontend/components/Modal/ModalInner/DeletePostModal/index.tsx
@@ -3,7 +3,7 @@ import { Button } from 'antd';
 import { useDispatch } from 'react-redux';
 import { modalHide } from '@/store/actions/modal';
 import { mutate } from 'swr';
-import { deletePost } from '@/store/actions/post';
+import { acceptPost, deletePost } from '@/store/actions/post';
 import { IPost } from '@/types/post';
 import { StyledDeletePostModal } from './styles';
 import { whereAreThePostLink } from '@/utils';
@@ -13,36 +13,54 @@ export interface IDeletePostModalProps {
     pageUserId?: number;
     postId: number;
     partyId?: number;
+    isOffer?: boolean;
+    isOfferConfirm?: boolean;
 }
 
 interface IDeletePostModal extends IDeletePostModalProps {
     setClose: () => void;
 }
 
-const DeletePostModal: React.FC<IDeletePostModal> = ({ pageUserId, postId, setClose, partyId }) => {
+const DeletePostModal: React.FC<IDeletePostModal> = ({
+    pageUserId,
+    postId,
+    setClose,
+    partyId,
+    isOffer,
+    isOfferConfirm,
+}) => {
     const dispatch = useDispatch();
 
     const deleteHanlder = () => {
-        const postUrl = whereAreThePostLink(pageUserId, partyId);
+        const postUrl = whereAreThePostLink(pageUserId, partyId, isOffer);
         deletePostMutate(postId, postUrl);
-        dispatch(deletePost(postId, postUrl));
+        if (isOfferConfirm && partyId) {
+            dispatch(acceptPost(partyId, postId, postUrl));
+        } else {
+            dispatch(deletePost(postId, postUrl));
+        }
         setClose();
     };
 
     return (
         <StyledDeletePostModal>
-            <h4>Вы дейтвительно хотите удалить пост?</h4>
-            <Button
-                type="primary"
-                danger
-                onClick={deleteHanlder}
-            >
-                Да
-            </Button>
+            {isOfferConfirm ? (
+                <>
+                    <h4>Вы дейтвительно хотите принять пост?</h4>
+                    <Button type="primary" onClick={deleteHanlder}>
+                        Да
+                    </Button>
+                </>
+            ) : (
+                <>
+                    <h4>Вы дейтвительно хотите удалить пост?</h4>
+                    <Button type="primary" danger onClick={deleteHanlder}>
+                        Да
+                    </Button>
+                </>
+            )}
         </StyledDeletePostModal>
     );
 };
 
 export default DeletePostModal;
-
-
